refactor(clawdstrike): extract guards and ruleset from policy.rs
Pure relocation:
- guards.rs: impl Policy::create_guards + pub(crate) PolicyGuards +
  builtin_guards_in_order (stable evaluation order preserved verbatim)
- ruleset.rs: RuleSet + impl (yaml_by_name/by_name/list). include_str! paths stay
  ../../rulesets/ (ruleset.rs is at the same depth as mod.rs); test_rulesets and
  test_rulesets_parse_validate_and_match_disk_registry pass.

mod.rs re-exports RuleSet (pub) and PolicyGuards (pub(crate), consumed by engine.rs).
Now-unused serde + crate::guards imports dropped from mod.rs. No logic edits.
